requirements.txt
python-dotenv==0.19.2
noise==1.2.2